// frontend/src/app/shared/enums/event-type.ts
export enum EventType {
  CHAMPION_KILL = "CHAMPION_KILL",
  ELITE_MONSTER_KILL = "ELITE_MONSTER_KILL",
  BUILDING_KILL = "BUILDING_KILL",
  TURRET_PLATE_DESTROYED = "TURRET_PLATE_DESTROYED",
  ITEM_PURCHASED = "ITEM_PURCHASED",
  ITEM_UNDO = "ITEM_UNDO",
  ITEM_SOLD = "ITEM_SOLD",
  ITEM_DESTROYED = "ITEM_DESTROYED",
  GAME_END = "GAME_END",
}