// frontend/src/app/shared/enums/monster-type.ts
export enum MonsterType {
  DRAGON = "DRAGON",
  RIFTHERALD = "RIFTHERALD",
  BARON_NASHOR = "BARON_NASHOR",
  HORDE = "HORDE",
}

export enum MonsterSubType {
  EARTH_DRAGON = "EARTH_DRAGON",
  WATER_DRAGON = "WATER_DRAGON",
  FIRE_DRAGON = "FIRE_DRAGON",
  AIR_DRAGON = "AIR_DRAGON",
  HEXTECH_DRAGON = "HEXTECH_DRAGON",
  CHEMTECH_DRAGON = "CHEMTECH_DRAGON",
  ELDER_DRAGON = "ELDER_DRAGON",
}

export enum DragonSoulType {
  AIR = "AIR",
  CHEMTECH = "CHEMTECH",
  EARTH = "EARTH",
  FIRE = "FIRE",
  HEXTECH = "HEXTECH",
  WATER = "WATER",
}